Consolidate to useConversationThread
